Refactor: moves db fixtures to with_dbs/conftest
