+++ services/catalog/tests/unit/with_dbs/conftest.py
@@ -3,11 +3,23 @@
 # pylint:disable=redefined-outer-name
 
 
+import itertools
+import random
+from typing import Any, Callable, Dict, Iterator, List, Tuple
+
 import pytest
 import sqlalchemy as sa
+from aiopg.sa.engine import Engine
+from faker import Faker
 from fastapi import FastAPI
+from simcore_postgres_database.models.products import products
 from simcore_service_catalog.api.dependencies.director import get_director_api
 from simcore_service_catalog.core.application import init_app
+from simcore_service_catalog.db.tables import (
+    groups,
+    services_access_rights,
+    services_meta_data,
+)
 from starlette.testclient import TestClient
 
 
@@ -35,3 +47,190 @@ async def director_mockup(loop, app: FastAPI):
     yield
 
     app.dependency_overrides[get_director_api] = None
+
+
+# DATABASE tables fixtures -----------------------------------
+#
+# These are the tables accessible by the catalog service:
+#
+# - services_meta_data
+#   -> groups(gid)@owner
+# - services_access_rights
+#   -> groups(gid)@owner,
+#   -> products(name)@produt_name
+#   -> services_meta_data@key, version
+# - services_consume_filetypes
+#
+# and therefore these are the coupled tables
+# - groups
+# - products
+#
+
+
+@pytest.fixture()
+async def products_names(aiopg_engine: Engine) -> Iterator[List[str]]:
+    """Inits products db table and returns product names"""
+    data = [
+        # already upon creation: ("osparc", r"([\.-]{0,1}osparc[\.-])"),
+        ("s4l", r"(^s4l[\.-])|(^sim4life\.)|(^api.s4l[\.-])|(^api.sim4life\.)"),
+        ("tis", r"(^tis[\.-])|(^ti-solutions\.)"),
+    ]
+
+    # pylint: disable=no-value-for-parameter
+
+    async with aiopg_engine.acquire() as conn:
+        # NOTE: The 'default' dialect with current database version settings does not support in-place multirow inserts
+        for name, regex in data:
+            stmt = products.insert().values(name=name, host_regex=regex)
+            await conn.execute(stmt)
+
+    names = [
+        "osparc",
+    ] + [items[0] for items in data]
+    yield names
+
+    async with aiopg_engine.acquire() as conn:
+        await conn.execute(products.delete())
+
+
+@pytest.fixture()
+async def user_groups_ids(aiopg_engine: Engine) -> Iterator[List[int]]:
+    """Inits groups table and returns group identifiers"""
+
+    cols = ("gid", "name", "description", "type", "thumbnail", "inclusion_rules")
+    data = [
+        (34, "john.smith", "primary group for user", "PRIMARY", None, {}),
+        (
+            20001,
+            "Team Black",
+            "External testers",
+            "STANDARD",
+            "http://mib.org",
+            {"email": "@(foo|testers|mib)+.(org|com)$"},
+        ),
+    ]
+    # pylint: disable=no-value-for-parameter
+
+    async with aiopg_engine.acquire() as conn:
+        for row in data:
+            # NOTE: The 'default' dialect with current database version settings does not support in-place multirow inserts
+            stmt = groups.insert().values(**dict(zip(cols, row)))
+            await conn.execute(stmt)
+
+    gids = [
+        1,
+    ] + [items[0] for items in data]
+
+    yield gids
+
+    async with aiopg_engine.acquire() as conn:
+        await conn.execute(services_meta_data.delete())
+        await conn.execute(groups.delete().where(groups.c.gid.in_(gids[1:])))
+
+
+@pytest.fixture()
+async def services_db_tables_injector(aiopg_engine: Engine) -> Callable:
+    """Returns a helper function to init
+    services_meta_data and services_access_rights tables
+    """
+    # pylint: disable=no-value-for-parameter
+
+    async def inject_in_db(fake_catalog: List[Tuple]):
+        # [(service, ar1, ...), (service2, ar1, ...) ]
+
+        async with aiopg_engine.acquire() as conn:
+            # NOTE: The 'default' dialect with current database version settings does not support in-place multirow inserts
+            for service in [items[0] for items in fake_catalog]:
+                stmt_meta = services_meta_data.insert().values(**service)
+                await conn.execute(stmt_meta)
+
+            for access_rights in itertools.chain(items[1:] for items in fake_catalog):
+                stmt_access = services_access_rights.insert().values(access_rights)
+                await conn.execute(stmt_access)
+
+    yield inject_in_db
+
+    async with aiopg_engine.acquire() as conn:
+        await conn.execute(services_access_rights.delete())
+        await conn.execute(services_meta_data.delete())
+
+
+@pytest.fixture()
+async def service_catalog_faker(
+    user_groups_ids: List[int],
+    products_names: List[str],
+    faker: Faker,
+) -> Callable:
+    """Returns a fake factory that creates catalog data
+    that can be used to fill services_meta_data and services_access_rights tables
+    """
+    everyone_gid, user_gid, team_gid = user_groups_ids
+
+    def _random_service(**overrides) -> Dict[str, Any]:
+        data = dict(
+            key=f"simcore/services/{random.choice(['dynamic', 'computational'])}/{faker.name()}",
+            version=".".join([str(faker.pyint()) for _ in range(3)]),
+            owner=user_gid,
+            name=faker.name(),
+            description=faker.sentence(),
+            thumbnail=random.choice([faker.image_url(), None]),
+            classifiers=[],
+            quality={},
+        )
+        data.update(overrides)
+        return data
+
+    def _random_access(service, **overrides) -> Dict[str, Any]:
+        data = dict(
+            key=service["key"],
+            version=service["version"],
+            gid=random.choice(user_groups_ids),
+            execute_access=faker.pybool(),
+            write_access=faker.pybool(),
+            product_name=random.choice(products_names),
+        )
+        data.update(overrides)
+        return data
+
+    def _fake_factory(
+        key, version, team_access=None, everyone_access=None, product=products_names[0]
+    ) -> Tuple[Dict[str, Any], ...]:
+
+        service = _random_service(key=key, version=version)
+
+        # owner always has full-access
+        owner_access = _random_access(
+            service,
+            gid=service["owner"],
+            execute_access=True,
+            write_access=True,
+            product_name=product,
+        )
+
+        fakes = [
+            service,
+            owner_access,
+        ]
+        if team_access:
+            fakes.append(
+                _random_access(
+                    service,
+                    gid=team_gid,
+                    execute_access="x" in team_access,
+                    write_access="w" in team_access,
+                    product_name=product,
+                )
+            )
+        if everyone_access:
+            fakes.append(
+                _random_access(
+                    service,
+                    gid=everyone_gid,
+                    execute_access="x" in everyone_access,
+                    write_access="w" in everyone_access,
+                    product_name=product,
+                )
+            )
+        return tuple(fakes)
+
+    return _fake_factory
